% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{AUC}
\alias{AUC}
\title{ROC area under curve}
\usage{
AUC(pred, obs)
}
\arguments{
\item{pred}{Numeric vector of predicted values}

\item{obs}{Numeric vector of observed values or factor with two levels}
}
\value{
Returns the AUC
}
\description{
Estimates AUC from predicted and observed values
}
\examples{
glm1 <- glm(case ~ spontaneous + age, data=infert, family="binomial")
pred1 <- fitted(glm1)
AUC(pred1, infert$case)
}
