% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{cat_summary}
\alias{cat_summary}
\title{Auxiliary functions for report.data.frame}
\usage{
cat_summary(x, digits)
}
\arguments{
\item{x}{A vector of class character}

\item{digits}{Number of digits for rounding}
}
\value{
Returns a summary of a categorical variable
}
\description{
Internal function for categorical summary
}
