% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Auxiliary functions for report.data.frame}
\usage{
numeric_summary(x, digits)
}
\arguments{
\item{x}{A vector of class numeric}

\item{digits}{Number of digits for rounding}
}
\value{
Returns a summary of a numerical variable
}
\description{
Internal function for numerical summary
}
