% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{report}
\alias{report}
\title{Generic function for creating reporting tables}
\usage{
report(x, ...)
}
\arguments{
\item{x}{An compatibleobject that can be summarized}

\item{...}{further arguments passed to make_table}
}
\value{
A data frame with the report table
}
\description{
Generic function for creating reporting tables.
}
\examples{
report(iris)  #Report of descriptive statistics
lm1 <- lm(Petal.Length ~ Sepal.Width + Species, data=iris)
report(lm1)   #Report of model
}
