% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_model.R
\name{RMSE}
\alias{RMSE}
\title{Root mean squared error}
\usage{
RMSE(pred, obs)
}
\arguments{
\item{pred}{Numeric vector of predicted values}

\item{obs}{Numeric vector of observed values}
}
\value{
Returns the RMSE
}
\description{
Estimates root mean squared error from predicted and observed values
}
\examples{
lm1 <- lm(Petal.Length ~ Sepal.Width + Species, data=iris)
pred1 <- fitted(lm1)
RMSE(pred1, iris$Petal.Length)
}
